replace int with double for capacity and engines, add KG to truck capacity to make it more clear

+++ Motorcycle.cs
@@ -3,12 +3,12 @@
 //motorcycle class
 public class Motorcycle : Vehicle
 {
-    public int EngineCapacity { get; private set; }
+    public double EngineCapacity { get; private set; }
     public string FuelType { get; private set; }
     public bool HasFairing { get; private set; }
 
     public Motorcycle(string model, string manufacturer, int year, double rentalPrice,
-        int enginecapacity, string fueltype, bool hasfairing) : base(model, manufacturer, year, rentalPrice)
+        double enginecapacity, string fueltype, bool hasfairing) : base(model, manufacturer, year, rentalPrice)
     {
         EngineCapacity = enginecapacity;
         FuelType = fueltype;
